refactor: split amp settings transport
